requirements.txt
apache-airflow
sqlalchemy
praw
python-dotenv